requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
pandas>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
drf-yasg>=1.21.0
pyyaml>=6.0.0
uritemplate>=4.1.1 